test: merge Prometheus name tests into existing MetricsHelperTests
The new tests collided with the existing MetricsHelperTests class (CS0101).
Fold them into that class instead of a duplicate file.
